New Update

+++ database/migrations/2024_04_22_122258_create_router_plans_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->longText('details');
             $table->string('modify_prices')->default('no');
             $table->string('price_alteration_option')->default('percentage');
-            $table->integer('percentage')->default(0);
-            $table->decimal('added_amount')->default(0);
+            $table->decimal('percentage', 20, 2)->default(0.00);
+            $table->decimal('added_amount', 20, 2)->default(0.00);
             $table->timestamps();
         });
     }
